feat: rename nginx-ingress filename
